Conflicts:
	rftk-native/buffers/BufferCollection.cpp
	rftk-native/buffers/BufferCollection.h
	rftk-native/buffers/setup.py
	rftk-native/buildall.sh
	rftk-native/feature_extractors/AxisAlignedFeatureExtractor.cpp
	rftk-native/feature_extractors/AxisAlignedFeatureExtractor.h
	rftk-native/feature_extractors/RandomProjectionFeatureExtractor.cpp

+++ rftk-native/buffers/MatrixBuffer.h
@@ -129,7 +129,7 @@ MatrixBufferTemplate<T>::~MatrixBufferTemplate()
 template <class T>
 void MatrixBufferTemplate<T>::Resize(int m, int n)
 {
-    if(m*n > mData.size())
+    if(static_cast<size_t>(m*n) > mData.size())
     {
         mData.resize(m*n);
     }
